DefaultFileSuggestionService to correctly resolve absolute paths of files in groups
I frequently ran into the assertion at ChatInputViewModel.swift#L546 while trying out cmd with a project I'm working on. It turned out that the paths that belong to FileSuggestions returned by DefaultFileSuggestionService were invalid for many files. They all missed the sub-dirs inside the project's root.

After some more digging, I found out that PBXFileReference's path is relative to the project's root dir only if the file is organized in a buildable Folder (introduced in Xcode 16). If a file is within a Group, the file ref's path only contains the file name. So we need to walk up the hierarchy all the way to the root dir to construct the relative path.

This PR does that, and updates tests to prevent this issue from happening again in the future.