Update shfmt cmd in the pre-commit githook to correctly search for all shell files
The current "shfmt -w ./**/*.sh" cmd failed consistently on my machine. The error message was "lstat ./**/*.sh: no such file or directory". I think it searched for a file with that name, literally.

Fix by giving shfmt the current dir. This should work because its documentary states that "If a given path is a
directory, all shell scripts found under that directory will be used". Also tested with "shfmt -f ." and all shell files in the project were listed.